# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "NyanPass 高性能中转面板 API 文档",
        "title": "NyanPass API",
        "termsOfService": "https://nyanpass.io/terms",
        "contact": {
            "name": "API Support",
            "url": "https://nyanpass.io/support",
            "email": "support@nyanpass.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/admin/nodes": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Node"
                ],
                "summary": "添加节点（管理员）",
                "parameters": [
                    {
                        "description": "节点信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.AddNodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/nodes/{id}": {
            "get": {
                "tags": [
                    "Admin/Node"
                ],
                "summary": "获取节点详情（管理员）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "节点ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Node"
                ],
                "summary": "更新节点（管理员）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "节点ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "节点信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.UpdateNodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admin/Node"
                ],
                "summary": "删除节点（管理员）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "节点ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/admin/nodes/{id}/refresh": {
            "post": {
                "tags": [
                    "Admin/Node"
                ],
                "summary": "刷新节点状态（管理员）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "节点ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/orders": {
            "get": {
                "tags": [
                    "Admin/Order"
                ],
                "summary": "获取订单列表（管理员）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "订单状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/orders/{id}": {
            "get": {
                "tags": [
                    "Admin/Order"
                ],
                "summary": "获取订单详情（管理员）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "订单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admin/Order"
                ],
                "summary": "删除订单（管理员）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "订单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/orders/{id}/paid": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Order"
                ],
                "summary": "标记订单已支付（管理员）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "订单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "支付方式",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.MarkPaidRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/orders/{id}/refund": {
            "post": {
                "tags": [
                    "Admin/Order"
                ],
                "summary": "退款（管理员）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "订单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/user/orders": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "获取我的订单列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "创建订单",
                "parameters": [
                    {
                        "description": "订单信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.CreateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/user/orders/{id}": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "获取订单详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "订单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/user/orders/{id}/cancel": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "取消订单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "订单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.MarkPaidRequest": {
            "type": "object",
            "required": [
                "pay_method"
            ],
            "properties": {
                "pay_method": {
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务码",
                    "type": "integer"
                },
                "data": {
                    "description": "数据"
                },
                "msg": {
                    "description": "提示信息",
                    "type": "string"
                }
            }
        },
        "service.AddNodeRequest": {
            "type": "object",
            "required": [
                "address",
                "api_port",
                "api_token",
                "name"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "api_port": {
                    "type": "integer"
                },
                "api_token": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "rate": {
                    "type": "number",
                    "minimum": 0
                },
                "region": {
                    "type": "string"
                }
            }
        },
...
        "service.CreateOrderRequest": {
            "type": "object",
            "required": [
                "plan_id"
            ],
            "properties": {
                "coupon_id": {
                    "type": "integer"
                },
                "plan_id": {
                    "type": "integer"
                },
                "remark": {
                    "type": "string"
                }
            }
        },
...
        "service.UpdateNodeRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "api_port": {
                    "type": "integer"
                },
                "api_token": {
                    "type": "string"
                },
                "carrier": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "group_id": {
                    "type": "string"
                },
                "insecure": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "rate": {
                    "type": "number"
                },
                "region": {
                    "type": "string"
                },
                "sort": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Token 认证，格式：Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}